+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * fix(Link): define NuxtLinkProps instead of importing from `#app`
 * docs: update props schema to prevent hydration issues
 * docs(ComponentCode): Improve import generation
+* docs(md): Improve llms format
+* docs(md): Improve llms format callout
 
 ## [2.1.5](https://github.com/bitrix24/b24ui/compare/v2.1.4...v2.1.5) (2025-11-19)